Refacoring so that I can run both MockHpmsFetcher and the real implementation in the test envioronment.  The single test that requires the mock is now specifically configured for that.
