+++ hpms/src/test/java/gov/cms/ab2d/hpms/SpringBootTestApp.java
@@ -1,18 +1,14 @@
 package gov.cms.ab2d.hpms;
 
-import gov.cms.ab2d.hpms.service.HPMSFetcherImpl;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.FilterType;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"gov.cms.ab2d.common", "gov.cms.ab2d.hpms", "gov.cms.ab2d.eventlogger"},
-        excludeFilters = {
-            @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, value = { HPMSFetcherImpl.class }) })
+@ComponentScan(basePackages = {"gov.cms.ab2d.common", "gov.cms.ab2d.hpms", "gov.cms.ab2d.eventlogger"})
 @EntityScan(basePackages = {"gov.cms.ab2d.common.model"})
 @EnableJpaRepositories("gov.cms.ab2d.common.repository")
 @PropertySource("classpath:application.common.properties")
